noizu_scaffolding v1.1.5 Noizu.ERP protocol

Types

t()

t() :: term

Functions

entity(obj, options \\ %{})

Convert to scaffolding.struct object. Options may be passed to coordinate actions like expanding embedded references.

entity!(obj, options \\ %{})

Convert to scaffolding.struct object Options may be passed to coordinate actions like expanding embedded references. (With transaction wrapper if required)

record(obj, options \\ %{})

Convert to persistence object. Options may be passed to coordinate actions like expanding embedded references.

record!(obj, options \\ %{})

Convert to persistence object Options may be passed to coordinate actions like expanding embedded references. (With transaction wrapper if required)

ref(obj)

Cast to noizu reference object

sref(obj)

Cast to noizu string reference object