noizu_scaffolding v0.1.0 Noizu.Scaffolding.AuditEngineBehaviour behaviour

This Protocol provides a generic way for our scaffolding system to record audit worthy events. Actual implementation is left to the user but notes on a possible implemention are included in this projects README.md file.

Types

details()

details() :: any

entity()

entity() :: {:ref, atom, nmid} | any

event()

event() :: atom | tuple

nmid()

nmid() :: integer | atom | String.t

note()

note() :: nil | String.t

response()

response() :: :ok | {:error, details} | any

Callbacks

audit(event, details, entity, arg3, note)

audit(event, details, entity, Noizu.Scaffolding.CallingContext.t, note) :: response

Save event occurence to persistence layer with out wrapping within a Mnesia transaction.

audit!(event, details, entity, arg3, note)

audit!(event, details, entity, Noizu.Scaffolding.CallingContext.t, note) :: response

Save event occurence to persistence layer including mnesia transaction if mnesia backend used.