noizu_scaffolding v1.1.4 Noizu.Scaffolding.EntityBehaviour behaviour

This Behaviour provides some callbacks needed for the Noizu.ERP (EntityReferenceProtocol) to work smoothly.

Types

details()

details() :: any

entity_obj()

entity_obj() :: any

entity_record()

entity_record() :: any

entity_reference()

entity_reference ::
  entity_obj |
  entity_record |
  entity_tuple_reference |
  entity_string_reference

entity_string_reference()

entity_string_reference() :: String.t

entity_tuple_reference()

entity_tuple_reference() :: {:ref, module, nmid}

error()

error() :: {:error, details}

nmid()

nmid() :: integer | atom | String.t | tuple

options()

options() :: Map.t | nil

Callbacks

as_record(entity_obj)

as_record(entity_obj) :: entity_record | error

Converts entity into record format. Aka extracts any fields used for indexing with the expected database table looking something like %Table{ identifier: entity.identifier, ... any_indexable_fields: entity.indexable_field, ... entity: entity } The default implementation assumes table structure if simply %Table{identifier: entity.identifier, entity: entity} therefore you will need to overide this implementation if you have any indexable fields. Future versions of the entity behaviour will accept an indexable field option that will insert expected fields and (if indicated) do simple type casting such as transforming DateTime.t fields into utc time stamps or {time_zone, year, month, day, hour, minute, second} tuples for efficient range querying.

entity(entity_reference, options)

entity(entity_reference, options) :: entity_obj | error

Returns entity, given an identifier, ref tuple, ref string or other known identifier type. Where an entity is a EntityBehaviour implementing struct.

entity!(entity_reference, options)

entity!(entity_reference, options) :: entity_obj | error

Returns entity, given an identifier, ref tuple, ref string or other known identifier type. Wrapping call in transaction if required. Where an entity is a EntityBehaviour implementing struct.

record(entity_reference, options)

record(entity_reference, options) :: entity_record | error

Returns record, given an identifier, ref tuple, ref string or other known identifier type. Where a record is the raw mnesia table entry, as opposed to a EntityBehaviour based struct object.

record!(entity_reference, options)

record!(entity_reference, options) :: entity_record | error

Returns record, given an identifier, ref tuple, ref string or other known identifier type. Wrapping call in transaction if required. Where a record is the raw mnesia table entry, as opposed to a EntityBehaviour based struct object.

ref(entity_reference)

ref(entity_reference) :: entity_tuple_reference | error

Returns appropriate {:ref|:ext_ref, module, identifier} reference tuple

sref(entity_reference)

sref(entity_reference) :: entity_string_reference | error

Returns appropriate string encoded ref. E.g. ref.user.1234

sref_module()

sref_module() :: String.t

Returns the string used for preparing sref format strings. E.g. a User struct might use the string "user" as it's sref_module resulting in sref strings like ref.user.1234.