noizu_scaffolding v1.1.7 Noizu.Scaffolding.UnsupportedModule

Types

t()

t() :: %Noizu.Scaffolding.UnsupportedModule{reference: any}

Functions

as_record(item)

Converts entity into record format. Aka extracts any fields used for indexing with the expected database table looking something like %Table{ identifier: entity.identifier, ... any_indexable_fields: entity.indexable_field, ... entity: entity } The default implementation assumes table structure if simply %Table{identifier: entity.identifier, entity: entity} therefore you will need to overide this implementation if you have any indexable fields. Future versions of the entity behaviour will accept an indexable field option that will insert expected fields and (if indicated) do simple type casting such as transforming DateTime.t fields into utc time stamps or {time_zone, year, month, day, hour, minute, second} tuples for efficient range querying.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.as_record/1.

entity(item, options)

Returns entity, given an identifier, ref tuple, ref string or other known identifier type. Where an entity is a EntityBehaviour implementing struct.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.entity/2.

entity!(item, options)

Returns entity, given an identifier, ref tuple, ref string or other known identifier type. Wrapping call in transaction if required. Where an entity is a EntityBehaviour implementing struct.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.entity!/2.

id(item)

Return identifier of ref, sref, entity or record

Callback implementation for Noizu.Scaffolding.EntityBehaviour.id/1.

record(item, options)

Returns record, given an identifier, ref tuple, ref string or other known identifier type. Where a record is the raw mnesia table entry, as opposed to a EntityBehaviour based struct object.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.record/2.

record!(item, options)

Returns record, given an identifier, ref tuple, ref string or other known identifier type. Wrapping call in transaction if required. Where a record is the raw mnesia table entry, as opposed to a EntityBehaviour based struct object.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.record!/2.

ref(item)

Returns appropriate {:ref|:ext_ref, module, identifier} reference tuple

Callback implementation for Noizu.Scaffolding.EntityBehaviour.ref/1.

sref(item)

Returns appropriate string encoded ref. E.g. ref.user.1234

Callback implementation for Noizu.Scaffolding.EntityBehaviour.sref/1.

sref_module()

Returns the string used for preparing sref format strings. E.g. a User struct might use the string "user" as it's sref_module resulting in sref strings like ref.user.1234.

Callback implementation for Noizu.Scaffolding.EntityBehaviour.sref_module/0.